requirements.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cloudinary==1.32.0
cryptography==40.0.2
dj-database-url==2.0.0
Django==4.2
django-cloudinary-storage==0.3.0
django-cors-headers==3.14.0
django-environ==0.10.0
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
gunicorn==20.1.0
idna==3.4
phonenumbers==8.13.11
Pillow==9.5.0
pipreqs==0.4.13
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
pytz==2023.3
requests==2.29.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
yarg==0.1.9